+++ src/parser.h
@@ -31,6 +31,7 @@ enum class ParserOption : unsigned char {
   SINGLE_EXPR = 1u << 0u,
   NEED_HERE_END = 1u << 1u,     // for line continuation checking
   COLLECT_SIGNATURE = 1u << 2u, // for signature help
+  IMPLICIT_BLOCK = 1u << 3u,
 };
 
 template <>